Node activation functions moved to the utilities to unify common access to them from multiple packages
